Update Home_Control_System.ino
An updated version of the microcontroller sketch now handling the new format for devices and their command, as well as handling multiple devices at once, within a single serial message. It has been setup to handle any number of devices, and can easily be modified for any future addition of devices and their commands.